% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/academiadates-package.R
\docType{package}
\name{academiadates-package}
\alias{academiadates}
\alias{academiadates-package}
\title{academiadates: Tidy-Handling of Academic Dates}
\description{
Extends `tsibble` by adding a class that supports three terms per year.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Steven314/academiadates}
  \item \url{https://steven314.github.io/academiadates/}
  \item Report bugs at \url{https://github.com/Steven314/academiadates/issues}
}

}
\author{
\strong{Maintainer}: Steven Carter \email{steven.carter314@gmail.com}

}
\keyword{internal}
