% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{academiadates-scales}
\alias{academiadates-scales}
\alias{scale_x_yeartrimester}
\alias{scale_y_yeartrimester}
\alias{scale_colour_yeartrimester}
\alias{scale_color_yeartrimester}
\alias{scale_alpha_yeartrimester}
\alias{scale_fill_yeartrimester}
\alias{scale_size_yeartrimester}
\title{academiadates scales for ggplot2}
\usage{
scale_x_yeartrimester(...)

scale_y_yeartrimester(...)

scale_colour_yeartrimester(...)

scale_color_yeartrimester(...)

scale_alpha_yeartrimester(...)

scale_fill_yeartrimester(...)

scale_size_yeartrimester(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2:scale_date]{ggplot2::scale_x_date()}}.}
}
\value{
A ggproto object inheriting from \code{Scale}
}
\description{
Defines ggplot2 scales for the \code{\link{yeartrimester}} index.
}
