% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-academiadates.R, R/yeartrimester.R
\name{academiadates-vctrs}
\alias{academiadates-vctrs}
\alias{vec_arith.yeartrimester}
\alias{obj_print_data.yeartrimester}
\title{Internal vctrs methods}
\usage{
\method{vec_arith}{yeartrimester}(op, x, y, ...)

\method{obj_print_data}{yeartrimester}(x, ...)
}
\description{
These methods are the extensions that allow tsibble objects to
work with vctrs.
}
\keyword{internal}
