% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeartrimester.R
\name{yeartrimester}
\alias{yeartrimester}
\alias{make_yeartrimester}
\alias{is_yeartrimester}
\alias{academic_year}
\title{Represent year-trimester}
\usage{
yeartrimester(x, academic_start = 1)

make_yeartrimester(year = 1970L, trimester = 1L, academic_start = 1)

is_yeartrimester(x)

academic_year(x)
}
\arguments{
\item{x}{A date.}

\item{academic_start}{A numeric indicating the starting month of the academic
year.}

\item{year, trimester}{A vector of numerics which give year and trimester
values.}
}
\value{
A year-trimester (\code{yeartrimester}) object.
}
\description{
Create or coerce using \code{yeartrimester()}.
}
\section{Display}{

Use \code{\link[=format]{format()}} to display \code{yeartrimester} objects in desired format.
}

\examples{

# `academic_year()` helps to extract academic year
y <- yeartrimester(as.Date("2020-06-01"), academic_start = 6)
academic_year(y)
lubridate::year(y) # calendar years
}
