% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-trimester.R
\name{trimester}
\alias{trimester}
\title{Get the academic trimester of a date-time}
\usage{
trimester(x, type = "ay_year trimester", academic_start = 1)
}
\arguments{
\item{x}{A date-time object.}

\item{type}{The format to be returned for the trimester. Can be "trimester",
"ay_year trimester" (the default, e.g. '2025 T1'), "ay_year.trimester"
(academic year with trimester in numeric form), "year_start/end".,
"date_last", or "date_first". These are similar to the options for
\code{\link[lubridate:quarter]{lubridate::quarter}}.}

\item{academic_start}{A integer from 1 to 12 indicating the month for the
start of the academic year.}
}
\value{
A numeric or character depnding on \code{type}.
}
\description{
Trimesters divide the year into thirds. This is largely based on
\code{\link[lubridate:quarter]{lubridate::quarter}}.
}
\examples{
trimester(lubridate::as_date("2024-08-24"), academic_start = 5)

trimester(lubridate::today(), academic_start = 5)

trimester(lubridate::today(), type = "trimester", academic_start = 5)
}
